BMD-6423: ip access scope
