+++ .env.dev
@@ -1,5 +1,6 @@
 # docker network ip can be fetched with `ifconfig | grep docker -A 10 | grep inet | grep -v inet6` usually it is 172.17.0.1
 SITE_ENVIRONMENT=docker
+WEBPACK_MODE=development
 
 BPC_APP_ID=console
 BPC_APP_KEY=something_long_and_random_and_longer_than_32_chars
