+++ client/src/company/companyPlaninstancesMasks.js
@@ -1,0 +1,105 @@
+import React from "react";
+
+export default class CompanyPlaninstancesMasks extends React.Component {
+    constructor(props) {
+        super(props);
+        this.validateEmailmask = this.validateEmailmask.bind(this);
+        this.addEmailmask = this.addEmailmask.bind(this);
+        this.onChangeAddMask = this.onChangeAddMask.bind(this);
+        this.state = {
+            validMask: null,
+            userInput: false,
+            maskTimer: null
+        };
+    }
+
+    onChangeAddMask() {
+        this.setState({userInput: true});
+
+        clearTimeout(this.state.maskTimer);
+        this.setState({maskTimer: setTimeout(this.validateEmailmask(this.addMaskInput.value), 2000)});
+    }
+
+    validateEmailmask(value) {
+        const atIndex = value.indexOf('@');
+        const dotIndex = value.indexOf('.');
+        const valid = value.length > 0 &&     // There is a value
+            atIndex > -1 &&                     // There is an @
+            (atIndex + 1) < dotIndex &&         // There is a domain
+            value.length - dotIndex > 1;        // There is a top-level domain
+        this.setState({validMask: valid});
+        return valid;
+    }
+
+    addEmailmask() {
+        const input = this.addMaskInput.value;
+        if (this.validateEmailmask(input)) {
+            this.props.addEmailmask(input).then(() => {
+                this.addMaskInput.value = '';
+                this.setState({userInput: false});
+            });
+        }
+    }
+
+
+    render() {
+        let userSearchFeedback = null;
+        let userSearchBoxClass = 'form-group has-feedback';
+        if (this.state.userInput) {
+            if (this.state.validMask) {
+                userSearchFeedback =
+                    <span className="glyphicon glyphicon-ok form-control-feedback" aria-hidden="true"></span>;
+                userSearchBoxClass += ' has-success'
+            } else {
+                userSearchFeedback =
+                    <span className="glyphicon glyphicon-warning-sign form-control-feedback" aria-hidden="true"></span>;
+                userSearchBoxClass += ' has-error'
+            }
+        }
+        const _masks = this.props.masks || [];
+        const masks = _masks.map((mask) => {
+            return (
+                <tr key={mask}>
+                    <td>
+                        {mask}
+                    </td>
+                    <td>
+                        <div style={{textAlign: 'right'}}>
+                            <button type="button" className='btn btn-xs btn-danger' onClick={this.props.removeEmailmask}
+                                    style={{minWidth: '30px'}}>
+                                <span className='glyphicon glyphicon-trash' aria-hidden="true"></span>
+                                <span>Remove</span>
+                            </button>
+                        </div>
+                    </td>
+                </tr>
+            );
+        });
+        return (
+            <table className="table table-striped" style={{marginTop: '25px'}}>
+                <div><strong>Email Masks</strong></div>
+                <tbody>
+                {masks}
+                <tr>
+                    <div className={userSearchBoxClass} style={{marginTop: '5px'}}>
+                        <input
+                            type="text"
+                            name="addMaskInput"
+                            onChange={this.onChangeAddMask}
+                            className="form-control input-sm"
+                            ref={(addMaskInput) => this.addMaskInput = addMaskInput}
+                        />
+                        {userSearchFeedback}
+                        <div style={{textAlign: 'right', marginTop: '3px'}}>
+                            <button type="button" className='btn btn-xs btn-success' onClick={this.addEmailmask}
+                                    style={{minWidth: '90px'}}>
+                                <span className='glyphicon glyphicon-plus' aria-hidden="true"></span> <span>Add</span>
+                            </button>
+                        </div>
+                    </div>
+                </tr>
+                </tbody>
+            </table>
+        );
+    }
+}
