+++ client/src/company/companyUsers.js
@@ -1,9 +1,8 @@
+import React from "react";
+import CompanyUser from "./companyUser";
+import * as Bpc from "../components/bpc";
 
-const React = require('react');
-const Link = require('react-router-dom').Link;
-const Bpc = require('./components/bpc');
-
-module.exports = class extends React.Component {
+export default class CompanyUsers extends React.Component {
   constructor(props){
     super(props);
     this.validateEmail = this.validateEmail.bind(this);
@@ -156,7 +155,7 @@ module.exports = class extends React.Component {
     users = showTruncatedUserList ? users.slice(0, maxUsersToShow) : users;
 
     // Using the props.removeUser instead of this.removeUser, since this component does not need to clear the input field
-    let items = users.map((user) => <User key={user.uid} user={user} removeUser={this.props.removeUser} searchUser={this.searchUser} />)
+    let items = users.map((user) => <CompanyUser key={user.uid} user={user} removeUser={this.props.removeUser} searchUser={this.searchUser} />)
 
     if(showTruncatedUserList) {
       items.push(
@@ -224,95 +223,6 @@ module.exports = class extends React.Component {
           </div>
         </div>
       </div>
-    );
-  }
-}
-
-
-class User extends React.Component {
-  constructor(props){
-    super(props);
-    this.getAddedByDetails = this.getAddedByDetails.bind(this);
-    this.showAddedByDetails = this.showAddedByDetails.bind(this);
-    this.state = {
-      foundUser: null,
-      addedByUser: null,
-      showAddedByDetails: false
-    };
-  }
-
-
-  getUserDetails() {
-    return Bpc.request(`/users?provider=gigya&id=${ encodeURIComponent(this.props.user.uid) }`)
-    .then((users) => {
-      const foundUser = users.length === 1 ? users[0] : null;
-      if(foundUser) {
-        this.setState({ foundUser });
-      } else {
-        // Waiting on the Gigya webhook
-        setTimeout(() => this.getUserDetails(), 3000);
-      }
-    });
-  }
-
-
-  showAddedByDetails() {
-    this.getAddedByDetails()
-    .then(() => this.setState({ showAddedByDetails: true }))
-  }
-
-
-  getAddedByDetails() {
-    if(this.props.user.addedBy && this.props.user.addedBy.user) {
-      return Bpc.request(`/users/${ encodeURIComponent(this.props.user.addedBy.user) }`)
-      .then(addedByUser => this.setState({ addedByUser }));
-    } else {
-      return Promise.resolve();
-    }
-  }
-
-
-  componentDidMount() {
-    this.getUserDetails(this.props.user.uid);
-  }
-
-
-  render() {
-
-    let item = <span>{ this.props.user.uid }</span>
-    const foundUser = this.state.foundUser;
-
-    if(foundUser) {
-      item = <Link to={`/users?search=${ foundUser.id }`}>{ foundUser.email }</Link>
-    }
-
-    let addedByDetails = '';
-    if(this.state.addedByUser) {
-      addedByDetails = this.state.addedByUser.email;
-    } else if(this.props.user.addedBy) {
-      addedByDetails = this.props.user.addedBy.user || this.props.user.addedBy.system;
-    } 
-
-    return (
-      <tr key={this.props.user.uid}>
-        <td>
-          <div style={{ marginTop: '3px', marginBottom: '3px' }}>
-            { item }
-            { this.state.showAddedByDetails
-              ? <div>
-                  <div><em><small>Added: {this.props.user.addedAt || ''}</small></em></div>
-                  <div><em><small>Added by: { addedByDetails }</small></em></div>
-                </div>
-              : <button type="button" className="btn btn-sm btn-link" style={{marginLeft: '10px', padding: '0px'}} onClick={this.showAddedByDetails}>Details</button>
-            }
-          </div>
-        </td>
-        <td style={{ textAlign: 'right'}}>
-          <button type="button" className='btn btn-xs btn-danger' onClick={this.props.removeUser.bind(this, this.props.user)} style={{ minWidth: '90px' }}>
-            <span className='glyphicon glyphicon-trash' aria-hidden="true"></span> <span>Remove</span>
-          </button>
-        </td>
-      </tr>
     );
   }
 }
