+++ client/src/main.js
@@ -10,7 +10,7 @@ const Users = require('./users');
 const ConsoleUsers = require('./consoleUsers');
 const Application = require('./application');
 const Tools = require('./tools');
-const Companies = require('./companies');
+const Companies = require('./companies').default;
 const Payments = require('./payments');
 const Customers = require('./customers');
 const GoogleLogin = require('react-google-login').default;
