+++ client/src/validators/validateIp.js
@@ -1,0 +1,5 @@
+export function validateIpv4(value) {
+    // Regex for IP incl. CIDR validation
+    const regex = RegExp('^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(\/([0-9]|[1-2][0-9]|3[0-2]))?$');
+    return regex.test(value);
+}
